回退 'Pull Request !188 : 修复代码生成模块删除时报错的bug'

+++ pig-visual/pig-codegen/src/main/java/com/pig4cloud/pig/codegen/controller/GenDsConfController.java
@@ -101,7 +101,7 @@ public class GenDsConfController {
 	 */
 	@SysLog("删除数据源表")
 	@DeleteMapping("/{id}")
-	public R<Boolean> removeById(@PathVariable Long id) {
+	public R<Boolean> removeById(@PathVariable Integer id) {
 		return R.ok(datasourceConfService.removeByDsId(id));
 	}
